requirements.txt
guli==0.1
keyboard==0.13.5
matplotlib==3.5.3
mysql_connector_repackaged==0.3.1
opencv_python==4.6.0.66
pandas==1.4.3
PySimpleGUI==4.60.3
pywin32==304
qrcode==7.3.1
requests==2.28.1
schedule==1.1.0
